Move `vctrs_arg` declarations to arg.h
