+++ src/vctrs.h
@@ -135,47 +135,9 @@ SEXP vec_unspecified(R_len_t n);
 bool vec_is_unspecified(SEXP x);
 
 
-// Argument tags -------------------------------------------------
-
-/**
- * Structure for argument tags
- *
- * Argument tags are used in error messages to provide information
- * about which elements of nested data structures (such as tibbles)
- * fail to match a given type. They are generated lazily by the `fill`
- * method in order to avoid any cost when there is no error.
- *
- * `vctrs_arg` is meant to be used with C-style inheritance. It should
- * be stored as the first member of derived structs. See
- * <https://www.python.org/dev/peps/pep-3123/>.
- *
- * @member parent The previously active argument tag.
- * @member fill Takes a pointer to self, which can be cast back to the
- *   original type containing the data, and a buffer to fill. If the
- *   buffer is too small according to the `remaining` argument,
- *   `fill()` must return a negative error value.
- */
-struct vctrs_arg {
-  struct vctrs_arg* parent;
-  r_ssize_t (*fill)(struct vctrs_arg* self, char* buf, r_ssize_t remaining);
-};
-
-struct vctrs_arg_wrapper {
-  struct vctrs_arg iface;
-  const char* arg;
-};
-
-struct vctrs_arg_counter {
-  struct vctrs_arg iface;
-  R_len_t* i;
-};
-
-struct vctrs_arg_wrapper new_vctrs_arg(struct vctrs_arg* parent, const char* arg);
-struct vctrs_arg_counter new_counter_arg(struct vctrs_arg* parent, R_len_t* i);
-SEXP vctrs_arg(struct vctrs_arg* arg);
-
-
 // Vector methods ------------------------------------------------
+
+#include "arg.h"
 
 SEXP vec_proxy(SEXP x);
 R_len_t vec_size(SEXP x);
